Added new Icons for Resource and Question
Added a handfull of designables(for general UIView interface editing and for buttons.

Added the Qpopup and RPopup nibs.

Added class declarations for a new PopupView with extensions to load from a nib.

Were working happily with a prior version but post merge everything is up in the air.

Will be working on resolving conflicts and making everything clean.

Here they are though.